Refactor out common part of otel test case so can use stuff we used for TruCustom for TruChain.
